Update model logic and Statistics naming/decl conventions

+++ client_app/app/src/main/java/com/spellingo/client_app/UpdateModel.kt
@@ -1,6 +1,9 @@
 package com.spellingo.client_app
 
 import android.app.Application
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
+import androidx.core.content.ContextCompat
 import org.json.JSONException
 import org.json.JSONObject
 
@@ -9,6 +12,11 @@ import org.json.JSONObject
  * @param application ApplicationContext for database creation
  */
 abstract class UpdateModel(private val application: Application) {
+    private val locale = "uk" //TODO replace with setting
+    private val connectivity = ContextCompat.getSystemService(
+        application.applicationContext,
+        ConnectivityManager::class.java
+    )
     protected val wordDb = WordDatabase.getInstance(application)
     protected val histDb = HistoryDatabase.getInstance(application)
 
@@ -16,18 +24,16 @@ abstract class UpdateModel(private val application: Application) {
      * Try to fetch words from server
      * @param limit number of words to request from server
      * @param locale English locale
-     * @return number of duplicate words received
+     * @return number of words fetched
      */
     protected suspend fun tryFetchWords(limit: Int, locale: String): Int {
+        if(limit <= 0) return 0
+
         val wordList = mutableListOf<Word>()
         val histList = mutableListOf<History>()
 
         // HTTP request
         val response = HttpRequest().getWords(limit, locale)
-
-        // Word and History database accessors
-        val wordDao = wordDb.wordDao()
-        val histDao = histDb.historyDao()
 
         //TODO error handling for empty fields (e.g. audio)
 
@@ -48,31 +54,87 @@ abstract class UpdateModel(private val application: Application) {
         }
         catch(e: JSONException) {
             System.err.println(e.toString())
-            return limit
+            return 0
         }
+
+        // Word and History database accessors
+        val wordDao = wordDb.wordDao()
+        val histDao = histDb.historyDao()
 
         // Get existing requested words from word history
-        val existing = histDao.getExisting(histList.map{it.id})
+        val histExisting = histDao.getExisting(wordList.map { it.id } ).map {
+            it.id
+        }.toHashSet()
 
-        // Using DAO properties, History ignores dup keys and Word replaces them.
-        // This correct behaviour since locale changes can make smarter logic tricky.
-
-        if(wordList.size > 0 && histList.size > 0) {
-            wordDao.insert(*wordList.toTypedArray())
-            histDao.insert(*histList.toTypedArray())
+        val cacheLocalMap = wordDao.getExisting(wordList.map { it.id }).associate {
+            it.id to it.locale
         }
 
-        return existing.size
+        // Remove fetched words if they're in cache with the same locale OR they are in history
+        wordList.removeAll {
+            if(cacheLocalMap.containsKey(it.id)) {
+                cacheLocalMap[it.id] == locale
+            }
+            else {
+                histExisting.contains(it.id)
+            }
+        }
+
+        // Using DAO properties, History ignores dup keys and Word replaces them.
+        // This means that words in cache of a different locale are allowed to be downloaded
+        if(wordList.isNotEmpty()) {
+            wordDao.insert(*wordList.toTypedArray())
+            histDao.insert(*wordList.map {
+                History(it.id, 0, 0)
+            }.toTypedArray())
+        }
+        println(wordList) // DEBUG
+
+        return wordList.size
     }
 
-    abstract suspend fun downloadWords()
-    abstract suspend fun purgeReusedWords()
+    /**
+     * Returns whether connection is metered
+     * @return True if and only if connection isn't metered (can download)
+     */
+    protected open fun canDownload(): Boolean {
+        if(connectivity == null) {
+            return false
+        }
+        val currentNetwork = connectivity.activeNetwork
+        val caps = connectivity.getNetworkCapabilities(currentNetwork)
+        return if(caps == null) {
+            // default to metered dumb check
+            !connectivity.isActiveNetworkMetered
+        }
+        else {
+            // Allow Wifi and Ethernet connections
+            caps.hasTransport(NetworkCapabilities.TRANSPORT_WIFI)
+                    || caps.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET)
+        }
+    }
+
+    /**
+     * Download words from server
+     * @param limit number of words to request from server
+     * @param locale English locale
+     * @return number of words downloaded
+     */
+    protected abstract suspend fun downloadWords(locale: String): Int
+
+    /**
+     * Purge some words from local cache
+     * @return number of words purged
+     */
+    protected abstract suspend fun purgeReusedWords(): Int
 
     /**
      * Fetch words from the server
      */
     suspend fun generateWords() {
-        purgeReusedWords()
-        downloadWords()
+        if(canDownload()) {
+            purgeReusedWords()
+            downloadWords(locale)
+        }
     }
 }
